VNS-003 VRTConverterProxy : validations on setPendingAdmin
Problem: During the setting of the new admin on the contract, the previous admin can be the same as the new admin.

Solution: The new admin should not be the same as a previous admin. Add a validation to assert the condition.
Added unit test coverage for the admin scenarios